Correção no banco. Integração do Endereço com EmpresaCliente. Correção no relacionamento de Prestador e Plano.

+++ src/main/java/com/perseus/conectapro/Controller/PlanoController.java
@@ -22,7 +22,7 @@ public class PlanoController {
 
     //Buscar plano por id
     @GetMapping("/{id}")
-    public Plano buscarPlanoPorId(@PathVariable int id) { return planoService.buscarPlanoPorId(id); }
+    public Plano buscarPlanoPorId(@PathVariable Long id) { return planoService.buscarPlanoPorId(id); }
 
     //Cadastrar plano
     @PostMapping("/registro")
@@ -33,13 +33,13 @@ public class PlanoController {
 
     //Alterar plano
     @PutMapping("/{id}")
-    public ResponseEntity<Plano> alterarPlano(@PathVariable int id, @RequestBody Plano plano) {
+    public ResponseEntity<Plano> alterarPlano(@PathVariable Long id, @RequestBody Plano plano) {
         Plano planoCriado = planoService.cadastrarPlano(plano);
         return ResponseEntity.status(HttpStatus.CREATED).body(planoCriado);
     }
 
     //Deletar plano
     @DeleteMapping("/{id}")
-    public void deletarPlano(@PathVariable int id) { planoService.deletarPlano(id); }
+    public void deletarPlano(@PathVariable Long id) { planoService.deletarPlano(id); }
 
 }
