+++ src/main/java/com/perseus/conectapro/DTO/PrestadorCreateDTO.java
@@ -16,6 +16,6 @@ public class PrestadorCreateDTO extends UsuarioCreateDTO{
     private String descPrestador;
     private List<String> especialidades;
     private StatusDisponibilidadeEnum statusDisponibilidade;
-    private int idPlano;
+    private Long idPlano;
 
 }
