+++ src/main/java/com/perseus/conectapro/Repository/PlanoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PlanoRepository extends JpaRepository<Plano, Integer> {
+public interface PlanoRepository extends JpaRepository<Plano, Long> {
 }
